requirements.txt
Flask==3.0.0
requests==2.31.0
python-metar==1.4.0
